pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore management system with supplier ordering and inventory tracking"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
psycopg2-binary = "^2.9"
django-environ = "^0.11.2"
django-parler = "^2.3"
django-autocomplete-light = "^3.11"
celery = "^5.3.4"
redis = "^5.0.1"
django-redis = "^5.4.0"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
flower = "^2.0.1"
gevent = "^24.2.1"
django-post-office = "^3.8"
WeasyPrint = "^60.0"
python-barcode = "^0.15.1"
Pillow = "^10.2.0"
django-tables2 = "^2.7.0"
django-filter = "^23.5"
openpyxl = "^3.1.2"
django-import-export = "^3.3.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2.0.0"
django-htmx = "^1.17.3"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3"
django-extensions = "^3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
